Quiz_BE/requirements.txt
django
djangorestframework
python-dotenv
psycopg2-binary
gunicorn
uvicorn
django-cors-headers
whitenoise